+++ src/main/java/wooteco/subway/domain/Line.java
@@ -1,0 +1,4 @@
+package wooteco.subway.domain;
+
+public class Line {
+}
