+++ src/main/java/wooteco/subway/domain/Section.java
@@ -1,0 +1,6 @@
+package wooteco.subway.domain;
+
+public class Section {
+
+
+}
