+++ src/main/java/wooteco/subway/station/StationController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
+import wooteco.subway.domain.Station;
 
 @RestController
 public class StationController {
